+++ backend/zerowaste/src/main/java/com/zerowaste/zerowaste/services/users/AuthenticateUserDTO.java
@@ -1,4 +1,4 @@
-package com.zerowaste.zerowaste.application.services.users;
+package com.zerowaste.zerowaste.services.users;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
