+++ backend/zerowaste/src/main/java/com/zerowaste/zerowaste/services/users/AuthenticateUserService.java
@@ -1,4 +1,4 @@
-package com.zerowaste.zerowaste.application.services.users;
+package com.zerowaste.zerowaste.services.users;
 
 import java.io.IOException;
 import java.time.Instant;
