+++ backend/zerowaste/src/main/java/com/zerowaste/zerowaste/services/users/exceptions/UserWithSameEmailAlreadyExistsException.java
@@ -1,4 +1,4 @@
-package com.zerowaste.zerowaste.application.services.users.exceptions;
+package com.zerowaste.zerowaste.services.users.exceptions;
 
 public class UserWithSameEmailAlreadyExistsException extends RuntimeException {
     public UserWithSameEmailAlreadyExistsException(String message) {
