+++ backend/zerowaste/src/test/java/com/zerowaste/zerowaste/application/services/users/RegisterUserServiceTest.java
@@ -18,9 +18,11 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.test.util.ReflectionTestUtils;
 
 import com.zerowaste.zerowaste.application.interfaces.UsersRepository;
-import com.zerowaste.zerowaste.application.services.users.exceptions.UserWithSameEmailAlreadyExistsException;
 import com.zerowaste.zerowaste.domain.entities.user.User;
 import com.zerowaste.zerowaste.domain.entities.user.UserRole;
+import com.zerowaste.zerowaste.services.users.RegisterUserDTO;
+import com.zerowaste.zerowaste.services.users.RegisterUserService;
+import com.zerowaste.zerowaste.services.users.exceptions.UserWithSameEmailAlreadyExistsException;
 
 @ExtendWith(MockitoExtension.class)
 public class RegisterUserServiceTest {
